// Portal.Portal.Web/src/app/shared/infrastructure/PortalHttpClient/model/createRoleRequest.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateRoleRequest { 
    addEmployeePermission?: boolean;
    viewEmployeesListPermission?: boolean;
    editPersonalInformationPermission?: boolean;
    editJobInfoPermission?: boolean;
    editEducationPermission?: boolean;
    editPayrollBasicPermission?: boolean;
    editEmploymentHistoryPermission?: boolean;
    editSkillsAndLanguagesPermission?: boolean;
    editOtherInfoPermission?: boolean;
    allTimeOffRequestsPermission?: boolean;
    allNonComliancesPermission?: boolean;
    viewAdminPanelPermission?: boolean;
    viewRolesPermission?: boolean;
    name: string;
    description?: string;
}